+++ integration-app/src/test/java/uk/gov/crowncommercial/esourcing/api/ApiExceptionHandlerIT.java
@@ -21,13 +21,13 @@ import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import uk.gov.crowncommercial.esourcing.mvc.ErrorResponse;
-import uk.gov.crowncommercial.esourcing.mvc.MvcConfiguration;
+import uk.gov.crowncommercial.esourcing.app.AppConfiguration;
+import uk.gov.crowncommercial.esourcing.app.ErrorResponse;
 import uk.gov.crowncommercial.esourcing.service.TenderApiService;
 
 @WebMvcTest(controllers = {TendersApiController.class})
 @AutoConfigureMockMvc
-@Import({MvcConfiguration.class})
+@Import({AppConfiguration.class})
 public class ApiExceptionHandlerIT {
 
   @Autowired
